tools: selftests - create a separate hotplug target for full range test
On some systems, hotplug tests can hang forever waiting for cpu and memory
to be ready to be offlined.

A special hot-plug target is created to run the full range of hotplug
tests.  In default mode, hotplug tests run in safe mode with a limited
scope.  In limited mode, cpu-hotplug test is run on a single cpu as
opposed to all hotplug capable cpus, and memory hotplug test is run on 2%
of hotplug capable memory instead of 10%.  In addition to the above
change, cpu-hotplug is changed to change processor affinity to cpu 0 so it
doesn't impact itself while the test runs.

Signed-off-by: Shuah Khan <shuah.kh@samsung.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Kees Cook <keescook@chromium.org>
Cc: Michael Ellerman <michael@ellerman.id.au>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>