mm: memcontrol: rewrite uncharge API fix - clear page->mapping in migration
The page->mapping reset in migration was conditionalized so that a later
uncharge could use PageAnon() to figure out the page type.  But after "mm:
memcontrol: rewrite uncharge API fix - double migration", the old page is
uncharged directly in mem_cgroup_migrate() and so this is no longer
necessary.

Once all fixups are folded into the "rewrite uncharge API" patch,
there will be no reason and no explanation for this change anymore, so
revert it.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@suse.cz>
Cc: Hugh Dickins <hughd@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>